+++ NoveltyProducer/Painter.py
@@ -5,11 +5,11 @@ import json
 import numpy as np
 import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
-from datetime import datetime
+import datetime
 # from IPython.core.debugger import set_trace
 
 class Painter:
-
+    MAX_DELAY = 1/(24*60*60) # 1 second
     MEMORY = 5/(24*60*60) # 5 seconds
     DATE_FORMAT = '%Y-%m-%dT%H:%M:%S.%f'
     DISPLAY_DATE_FORMAT = '%M:%S'
@@ -23,7 +23,9 @@ class Painter:
         self.ip = ip_
         self.port = port_
         self.topic = topic_
+        
         # internal definitions
+        self.frametimes = []
         # init plot
         plt.ion()
         self.fig, self.ax = plt.subplots()
@@ -35,18 +37,30 @@ class Painter:
         self.ax.grid()
         # figure events
         self.fig.canvas.mpl_connect('close_event', self._on_close_figure)
+        # update plot the first time
+        self._draw()
         # client
         self.data = {}
         self.client = mqtt.Client()
         self.client.on_connect = self._on_connect
         self.client.on_message = self._on_message
         self._establish_connection()
+        # start redrawing
+        self._run()
+        
+    def _run(self):
+        """Forever loop to update plot.
+        """
+        while True:
+            # update plot
+            self._draw()
         
     def _establish_connection(self):
         """Establish connection to mqtt broker
         """
         self.client.connect(self.ip, self.port, 60)
-        self.client.loop_forever()
+        self.client.loop_start()
+        #self.client.loop_forever()
 
     def _on_connect(self, client, userdata, flags, rc):
         """Define what to do when connection was established.
@@ -61,6 +75,11 @@ class Painter:
         sample = json.loads(msg.payload.decode())
         # get latest timestamp
         current_time = self._datestr2num(sample['timestamp'])
+        # is current timestamp older than MAX_DELAY?
+        if (self._datestr2num(datetime.datetime.now().isoformat()) - self.MAX_DELAY) > current_time:
+            # skip current sample
+            return
+        # get theoretically oldest valid time
         oldest_time = current_time - self.MEMORY
         # get all channel names
         current_names = [*sample.keys()]
@@ -91,10 +110,7 @@ class Painter:
             # assign new data
             self.channels[name]['x'] = validx
             self.channels[name]['y'] = validy
-            
-        # update plot
-        self._draw()
-        
+           
     def _remove_old_data(self):
         """Make sure old data is removed from datastore.
         """
@@ -114,14 +130,20 @@ class Painter:
         """Plot latest datapoints.
         """
         # loop over each channel and update plot
-        for chn in self.channels:
+        for chn in self.channels.copy():
             self.channels[chn]['line'].set_data(self.channels[chn]['x'], self.channels[chn]['y'])
+            # has channel no content anymore?
+            if len(self.channels[chn]['x']) == 0:
+                # remove channel
+                self._remove_channel(chn)
         # make sure limits are correct
         self.ax.relim()
         self.ax.autoscale_view()
         # flush it.
         self.fig.canvas.draw()
         self.fig.canvas.flush_events()
+        # print framerate
+        # self._get_framerate()
         
     def _on_close_figure(self, evt):
         """Triggered when plot is closed by the user. End connection.
@@ -131,7 +153,7 @@ class Painter:
     def _datestr2num(self, str_):
         """Convert datestring to num
         """
-        return mdates.date2num(datetime.strptime(str_, self.DATE_FORMAT))
+        return mdates.date2num(datetime.datetime.strptime(str_, self.DATE_FORMAT))
         
     def _add_channel(self, channel_):
         """Add new line to plot for a new channel
@@ -141,3 +163,24 @@ class Painter:
         self.channels[channel_]['x'] = []
         self.channels[channel_]['y'] = []
         self.ax.legend(loc=2)
+        
+    def _remove_channel(self, channel_):
+        """Remove channel if it runs out of scope / has no datapoints left.
+        """
+        # remove line from plot
+        self.ax.lines.remove(self.channels[channel_]['line'])
+        # remove channel from list of channels
+        self.channels.pop(channel_)
+        # refresh legend
+        self.ax.legend(loc=2)
+        
+    def _get_framerate(self):
+        """Compute the current framerate (frames in the last second)"""
+        # get current time
+        ts = datetime.datetime.now()
+        # add current frame time
+        self.frametimes.append(ts)
+        # update frames list
+        self.frametimes = [frame for frame in self.frametimes if frame >= ts - datetime.timedelta(seconds=1)]
+        # print result
+        print("\rFPS:%i" % len(self.frametimes), end="")